Move RouteOptionDrawer
